requirements.txt
flask
flask-sqlalchemy
google-generativeai
python-dotenv
werkzeug
gunicorn
psycopg2-binary
requests